pyproject.toml
[project]
name = "service-for-business"
version = "0.1.0"
description = ""
authors = [
    {name = "Kateryna Dementieva",email = "katerynadementieva@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
